Many improvements to PerfStats
Added several new counters related to reads, writes, and replication.
Also added infrastructure for printing out cluster stats (the method
printClusterStats isn't quite ready for prime time yet).

Also, changed Object::getValueOffset return value from uint16_t to
uint32_t (even though key data must be less than 64KB, key metadata
can push the offset beyond 16 bits).

+++ src/CRamCloud.cc
@@ -1,4 +1,4 @@
-/* Copyright (c) 2010-2014 Stanford University
+/* Copyright (c) 2010-2015 Stanford University
  *
  * Permission to use, copy, modify, and distribute this software for any
  * purpose with or without fee is hereby granted, provided that the above
@@ -792,7 +792,7 @@ rc_multiRead(struct rc_client* client,
                 continue;
             }
 
-            uint16_t value_offset;
+            uint32_t value_offset;
             bool retval = mReadHelper->value->getValueOffset(&value_offset);
             if (!retval) {
                 read_requests[i]->status = STATUS_INVALID_OBJECT;
